Reorganized project to put helper services in folder and focus on StudentEvaluationService and InstructorEvaluationService
