+++ backend/chromeApiServer/src/main/java/com/canvas/service/InstructorEvaluationService.java
@@ -1,0 +1,10 @@
+package com.canvas.service;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * This class handles the Instructor side of the evaluation process.
+ */
+@Service
+public class InstructorEvaluationService {
+}
