+++ backend/chromeApiServer/src/main/java/com/canvas/service/helperServices/CanvasClientService.java
@@ -1,4 +1,4 @@
-package com.canvas.service;
+package com.canvas.service.helperServices;
 
 import com.canvas.service.models.ExtensionUser;
 import com.fasterxml.jackson.databind.JsonNode;
