+++ backend/chromeApiServer/src/main/java/com/canvas/service/helperServices/FileService.java
@@ -1,4 +1,4 @@
-package com.canvas.service;
+package com.canvas.service.helperServices;
 
 import java.io.File;
 import java.io.FileOutputStream;
