+++ backend/chromeApiServer/src/main/java/com/canvas/service/helperServices/JSONParsingService.java
@@ -1,4 +1,4 @@
-package com.canvas.service;
+package com.canvas.service.helperServices;
 
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
